added internal middleware to cast values from query to object

+++ src/main/kotlin/commons/query/Ping.kt
@@ -1,3 +1,3 @@
 package commons.query
 
-data class Ping(val ping: Long)
+data class Ping(val ping: Byte)
